+++ src/Section.php
@@ -4,18 +4,18 @@ declare(strict_types=1);
 
 namespace Bakame\HtmlTable;
 
-enum Section: string
+enum Section
 {
-    case Header = 'thead';
-    case Body = 'tbody';
-    case Footer = 'tfoot';
-    case None = '';
+    case thead;
+    case tbody;
+    case tfoot;
+    case none;
 
     public function xpath(): string
     {
         return match ($this) {
-            self::None => '//table/tr',
-            default => '//table/'.$this->value.'/tr',
+            self::none => '//table/tr',
+            default => '//table/'.$this->name.'/tr',
         };
     }
 }